Brings main's kids profiles, library-update coordinator, player action
notices, the web player's reorganised modules and the Android UI package
split into the Android web-parity work, at version 0.53.0.

The branch's flat UI files move into main's ui.catalog, ui.player and
ui.profile packages. Where main only reformatted a file the branch's
version is kept; where main changed behaviour — kids profiles in the
picker and on the player's Kids mark, action notices, error handling for
detail and poster lookups, update-library through the coordinator — that
change is carried into the branch's structure.

The state schema keeps main's step 3 (profiles.kids) and makes the
branch's preferences table step 4. A state file an earlier build left at
version 3 without the kids column gains it on open.

+++ crates/mediagram-core/src/shows/mod.rs
@@ -47,7 +47,7 @@ pub fn upsert(conn: &Connection, row: &TitleDetailsRow) -> rusqlite::Result<()> 
         params![
             SOURCE,
             kind_key(row.kind),
-            row.id as i64,
+            row.id,
             row.lang,
             row.overview,
             row.tagline,
@@ -75,7 +75,7 @@ pub fn get(conn: &Connection, kind: Kind, id: u64) -> rusqlite::Result<Option<Ti
                     total_seasons, total_episodes, {certification}
                FROM shows WHERE source = ?1 AND kind = ?2 AND id = ?3"
         ),
-        params![SOURCE, kind_key(kind), id as i64],
+        params![SOURCE, kind_key(kind), id],
         |row| {
             Ok(TitleDetailsRow {
                 kind,
